[~] Can now create slide bar texture and fixed int/float confusion

+++ include/rpg.h
@@ -27,7 +27,7 @@ settings *init_settings_struct(void);
 void data_settings(screen *screen1);
 button *create_button_data(char *text, scene *s, sfVector2f pos);
 text *search_for_text(scene *s, char *name);
-sfVector2f create_vector2f(int x, int y);
+sfVector2f create_vector2f(float x, float y);
 sfVector2f get_w_scale(scene *d, game_obj *);
 void event_manager(scene *d, sfEvent event);
 scene *get_scene(scene *d, state state);
